fix bug progress bar quota user

+++ src/main/java/fr/ippon/tatami/web/rest/AttachmentController.java
@@ -69,13 +69,13 @@ public class AttachmentController {
     }
 
     /**
-     *  GET /attachment/quota -> get quota for the domain
+     *  GET /attachment/quota -> get quota in % for the domain
      */
     @RequestMapping(value = "/rest/attachments/quota",
             method = RequestMethod.GET,
             produces = "application/json")
     @ResponseBody
-    public Collection<Long> getDomainQuota(){
+    public Collection<Integer> getDomainQuota(){
         return attachmentService.getDomainQuota();
     }
